requirements.txt
Django>=5.1,<5.2
pillow
mysqlclient
cloudinary
python-decouple